Split force logic into modules
