+++ src/dynamics/rigid_body/forces/system_param.rs
@@ -1,16 +1,5 @@
-//! Forces, torques, linear impulses, and angular impulses
-//! that can be applied to dynamic rigid bodies.
-
-#![allow(missing_docs)]
-
 use crate::{
-    dynamics::{
-        integrator::{self, IntegrationSet, VelocityIntegrationData},
-        solver::{
-            solver_body::{SolverBody, SolverBodyInertia},
-            SolverDiagnostics,
-        },
-    },
+    dynamics::{integrator::VelocityIntegrationData, solver::solver_body::SolverBodyInertia},
     prelude::*,
 };
 use bevy::{
@@ -25,21 +14,7 @@ use bevy::{
     prelude::*,
 };
 
-#[cfg(feature = "2d")]
-pub(crate) type Torque = Scalar;
-
-#[cfg(feature = "3d")]
-pub(crate) type Torque = Vector;
-
-#[cfg(feature = "2d")]
-pub(crate) trait FloatZero {
-    const ZERO: Self;
-}
-
-#[cfg(feature = "2d")]
-impl FloatZero for Scalar {
-    const ZERO: Self = 0.0;
-}
+use super::{AccumulatedLocalAcceleration, AccumulatedLocalForces, AccumulatedWorldForces};
 
 #[derive(QueryData)]
 #[query_data(mutable)]
@@ -509,345 +484,4 @@ impl EntityForces<'_> {
             }
         }
     }
-}
-
-// TODO: Should accumulated forces and accelerations be `SparseSet` components?
-
-/// A component with the user-applied world forces and torques
-/// accumulated for a rigid body before the physics step.
-///
-/// Only entities with non-zero world forces have this component.
-/// It is added and removed automatically to reduce unnecessary work.
-#[derive(Component, Clone, Debug, Default, PartialEq, Reflect)]
-#[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
-#[cfg_attr(feature = "serialize", reflect(Serialize, Deserialize))]
-#[reflect(Component, Debug, Default, PartialEq)]
-pub struct AccumulatedWorldForces {
-    pub force: Vector,
-    pub torque: Torque,
-}
-
-/// A component with the user-applied local forces and torques
-/// accumulated for a rigid body before the physics step.
-///
-/// Only entities with non-zero local forces have this component.
-/// It is added and removed automatically to reduce unnecessary work.
-#[derive(Component, Clone, Debug, Default, PartialEq, Reflect)]
-#[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
-#[cfg_attr(feature = "serialize", reflect(Serialize, Deserialize))]
-#[reflect(Component, Debug, Default, PartialEq)]
-pub struct AccumulatedLocalForces {
-    pub force: Vector,
-    pub torque: Torque,
-}
-
-/// A component with the user-applied local acceleration
-/// accumulated for a rigid body before the physics step.
-///
-/// Only entities with non-zero local acceleration have this component.
-/// It is added and removed automatically to reduce unnecessary work.
-#[derive(Component, Clone, Debug, Default, PartialEq, Reflect)]
-#[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
-#[cfg_attr(feature = "serialize", reflect(Serialize, Deserialize))]
-#[reflect(Component, Debug, Default, PartialEq)]
-pub struct AccumulatedLocalAcceleration {
-    pub linear: Vector,
-    #[cfg(feature = "3d")]
-    pub angular: Vector,
-}
-
-pub struct ForcePlugin;
-
-impl Plugin for ForcePlugin {
-    fn build(&self, app: &mut App) {
-        app.register_type::<(
-            AccumulatedWorldForces,
-            AccumulatedLocalForces,
-            AccumulatedLocalAcceleration,
-        )>();
-
-        app.configure_sets(
-            PhysicsSchedule,
-            (
-                ForceSet::ApplyWorldForces
-                    .in_set(IntegrationSet::UpdateVelocityIncrements)
-                    .before(integrator::pre_process_velocity_increments),
-                ForceSet::Clear.in_set(SolverSet::PostSubstep),
-            ),
-        );
-
-        app.configure_sets(
-            SubstepSchedule,
-            ForceSet::ApplyLocalForces
-                .in_set(IntegrationSet::Velocity)
-                .before(integrator::integrate_velocities),
-        );
-
-        app.add_systems(
-            PhysicsSchedule,
-            apply_world_forces.in_set(ForceSet::ApplyWorldForces),
-        );
-        app.add_systems(
-            SubstepSchedule,
-            (apply_local_forces, apply_local_acceleration)
-                .chain()
-                .in_set(ForceSet::ApplyLocalForces),
-        );
-
-        app.add_systems(
-            PhysicsSchedule,
-            (
-                clear_accumulated_world_forces,
-                clear_accumulated_local_forces,
-                clear_accumulated_local_acceleration,
-            )
-                .in_set(ForceSet::Clear),
-        );
-    }
-}
-
-#[derive(SystemSet, Clone, Copy, Debug, PartialEq, Eq, Hash)]
-pub enum ForceSet {
-    ApplyWorldForces,
-    ApplyLocalForces,
-    Clear,
-}
-
-/// Clears [`AccumulatedWorldForces`] for all rigid bodies.
-///
-/// Continuously applied forces and torques are only reset to zero,
-/// while forces and torques that were already zero for an entire time step
-/// are removed from the entities.
-fn clear_accumulated_world_forces(
-    mut commands: Commands,
-    mut query: Query<(Entity, &mut AccumulatedWorldForces)>,
-) {
-    // Initialize a buffer for entities to remove the component from.
-    let mut entity_buffer = Vec::new();
-
-    for (entity, mut forces) in &mut query {
-        if forces.force != Vector::ZERO || forces.torque != Torque::ZERO {
-            // The force or torque was not zero, so these may be continuously applied forces.
-            // Just reset the forces and keep the component.
-            forces.force = Vector::ZERO;
-            forces.torque = Torque::ZERO;
-        } else {
-            // No forces or torques were applied for an entire time step, so we can remove the component.
-            entity_buffer.push(entity);
-        }
-    }
-
-    if entity_buffer.is_empty() {
-        return;
-    }
-
-    // Remove the component from all entities that had no forces or torques applied.
-    commands.queue(|world: &mut World| {
-        entity_buffer.into_iter().for_each(|entity| {
-            world.entity_mut(entity).remove::<AccumulatedWorldForces>();
-        });
-    });
-}
-
-/// Clears [`AccumulatedLocalForces`] for all rigid bodies.
-///
-/// Continuously applied forces and torques are only reset to zero,
-/// while forces and torques that were already zero for an entire time step
-/// are removed from the entities.
-fn clear_accumulated_local_forces(
-    mut commands: Commands,
-    mut query: Query<(Entity, &mut AccumulatedLocalForces)>,
-) {
-    // Initialize a buffer for entities to remove the component from.
-    let mut entity_buffer = Vec::new();
-
-    for (entity, mut forces) in &mut query {
-        if forces.force != Vector::ZERO || forces.torque != Torque::ZERO {
-            // The force or torque was not zero, so these may be continuously applied forces.
-            // Just reset the forces and keep the component.
-            forces.force = Vector::ZERO;
-            forces.torque = Torque::ZERO;
-        } else {
-            // No forces or torques were applied for an entire time step, so we can remove the component.
-            entity_buffer.push(entity);
-        }
-    }
-
-    if entity_buffer.is_empty() {
-        return;
-    }
-
-    // Remove the component from all entities that had no forces or torques applied.
-    commands.queue(|world: &mut World| {
-        entity_buffer.into_iter().for_each(|entity| {
-            world.entity_mut(entity).remove::<AccumulatedLocalForces>();
-        });
-    });
-}
-
-fn clear_accumulated_local_acceleration(
-    mut commands: Commands,
-    mut query: Query<(Entity, &mut AccumulatedLocalAcceleration)>,
-) {
-    // Initialize a buffer for entities to remove the component from.
-    let mut entity_buffer = Vec::new();
-
-    for (entity, mut acceleration) in &mut query {
-        #[cfg(feature = "2d")]
-        let non_zero_acceleration = acceleration.linear != Vector::ZERO;
-        #[cfg(feature = "3d")]
-        let non_zero_acceleration =
-            acceleration.linear != Vector::ZERO || acceleration.angular != Vector::ZERO;
-        if non_zero_acceleration {
-            // The acceleration was not zero, so this may be a continuously applied acceleration.
-            // Just reset the acceleration and keep the component.
-            acceleration.linear = Vector::ZERO;
-            #[cfg(feature = "3d")]
-            {
-                acceleration.angular = Vector::ZERO;
-            }
-        } else {
-            // No acceleration was applied for an entire time step, so we can remove the component.
-            entity_buffer.push(entity);
-        }
-    }
-
-    if entity_buffer.is_empty() {
-        return;
-    }
-
-    // Remove the component from all entities that had no acceleration applied.
-    commands.queue(|world: &mut World| {
-        entity_buffer.into_iter().for_each(|entity| {
-            world
-                .entity_mut(entity)
-                .remove::<AccumulatedLocalAcceleration>();
-        });
-    });
-}
-
-/// Applies gravity and locked axes to the linear and angular velocity increments of bodies.
-fn apply_world_forces(
-    mut bodies: Query<(
-        &mut VelocityIntegrationData,
-        &AccumulatedWorldForces,
-        &SolverBodyInertia,
-    )>,
-    mut diagnostics: ResMut<SolverDiagnostics>,
-) {
-    let start = crate::utils::Instant::now();
-
-    // TODO: Do we want to skip kinematic bodies here?
-    bodies
-        .par_iter_mut()
-        .for_each(|(mut integration, forces, mass_props)| {
-            // NOTE: The velocity increments are treated as accelerations at this point.
-
-            // Apply external forces and torques.
-            // NOTE: We ignore changes in the inertia tensor, keeping angular acceleration constant across substeps.
-            //       This may not be entirely correct, but is generally acceptable for world-space torque.
-            integration.linear_increment += mass_props.effective_inv_mass() * forces.force;
-            integration.angular_increment +=
-                mass_props.effective_inv_angular_inertia() * forces.torque;
-
-            // The `IntegrationPlugin` will take care of applying the time step
-            // and locked axes to the velocity increments.
-        });
-
-    diagnostics.update_velocity_increments += start.elapsed();
-}
-
-/// Applies [`AccumulatedLocalForces`] to the linear and angular velocity of bodies.
-fn apply_local_forces(
-    mut bodies: Query<
-        (
-            &mut SolverBody,
-            &AccumulatedLocalForces,
-            &Rotation,
-            &SolverBodyInertia,
-            Option<&LockedAxes>,
-        ),
-        With<SolverBody>,
-    >,
-    time: Res<Time<Substeps>>,
-    mut diagnostics: ResMut<SolverDiagnostics>,
-) {
-    let start = crate::utils::Instant::now();
-
-    let delta_secs = time.delta_secs_f64() as Scalar;
-
-    bodies
-        .par_iter_mut()
-        .for_each(|(mut body, forces, rotation, mass_props, locked_axes)| {
-            let rotation = body.delta_rotation * *rotation;
-            let locked_axes = locked_axes.map_or(LockedAxes::default(), |locked_axes| *locked_axes);
-
-            // NOTE:
-            //
-            // We could have a `LocalVelocityIncrements` component and apply these forces and torques
-            // to that once per time step rather than once per substep. However, that would have two caveats:
-            //
-            // 1. We need to store and manage both `AccumulatedLocalForces` and `LocalVelocityIncrements`
-            //    when local forces are applied.
-            // 2. Changes in the inertia tensor during substeps would not be considered.
-            //    (though we currently accept this for world-space torque)
-
-            // Compute the world-space accelerations with locked axes applied.
-            let linear_acceleration = locked_axes
-                .apply_to_vec(mass_props.effective_inv_mass() * (rotation * forces.force));
-            #[cfg(feature = "3d")]
-            let angular_acceleration = locked_axes.apply_to_angular_velocity(
-                mass_props.effective_inv_angular_inertia() * (rotation * forces.torque),
-            );
-
-            // Apply external forces and torques.
-            body.linear_velocity += linear_acceleration * delta_secs;
-            #[cfg(feature = "3d")]
-            {
-                body.angular_velocity += angular_acceleration * delta_secs;
-            }
-        });
-
-    diagnostics.integrate_velocities += start.elapsed();
-}
-
-/// Applies [`AccumulatedLocalAcceleration`] to the linear and angular velocity of bodies.
-///
-/// This should run in the substepping loop, just before [`IntegrationSet::Velocity`].
-fn apply_local_acceleration(
-    mut bodies: Query<(
-        &mut SolverBody,
-        &AccumulatedLocalAcceleration,
-        &Rotation,
-        Option<&LockedAxes>,
-    )>,
-    mut diagnostics: ResMut<SolverDiagnostics>,
-    time: Res<Time<Substeps>>,
-) {
-    let start = crate::utils::Instant::now();
-
-    let delta_secs = time.delta_secs_f64() as Scalar;
-
-    bodies
-        .par_iter_mut()
-        .for_each(|(mut body, acceleration, rotation, locked_axes)| {
-            let rotation = body.delta_rotation * *rotation;
-            let locked_axes = locked_axes.map_or(LockedAxes::default(), |locked_axes| *locked_axes);
-
-            // Compute the world space velocity increments with locked axes applied.
-            let world_linear_acceleration =
-                locked_axes.apply_to_vec(rotation * acceleration.linear);
-            #[cfg(feature = "3d")]
-            let world_angular_acceleration =
-                locked_axes.apply_to_vec(rotation * acceleration.angular);
-
-            // Apply acceleration.
-            body.linear_velocity += world_linear_acceleration * delta_secs;
-            #[cfg(feature = "3d")]
-            {
-                body.angular_velocity += world_angular_acceleration * delta_secs;
-            }
-        });
-
-    diagnostics.integrate_velocities += start.elapsed();
 }
